Moved to namespace, BinaryReader abstract again
I wanted to have BinaryReader as a non-abstract class so users may
specify BinaryReader and not worry about which implementation was
chosen. Unfortunately the C++ compiler hates non-abstract
non-implemented virtual methods, so I had to revert. To make things
easier as a library, I moved the classes to their own namespace.